+++ internal/provider/provider.go
@@ -844,6 +844,7 @@ func (p *IosxeProvider) DataSources(_ context.Context) []func() datasource.DataS
 func (p *IosxeProvider) Actions(_ context.Context) []func() action.Action {
 	return []func() action.Action{
 		NewCommitAction,
+		NewSaveConfigAction,
 	}
 }