+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `icmp_echo_frequency` attribute to `iosxe_sla` resource and data source
 - Add `iosxe_commit` action
+- Add `iosxe_save_config` action
 
 ## 0.13.0